simulator/requirements.txt
# simulator/requirements.txt
web3
python-dotenv
pyyaml
requests